Found the problem. I wasnt building the gpu stencils correctly I was using stencil_.size() instead of the stencil_size = stencil_[0].size()

+++ src/rbffd/derivative_cl.h
@@ -10,7 +10,7 @@
 
 class DerivativeCL : public Derivative, public CLBaseClass
 {
-    int total_num_stencil_elements; 
+    size_t total_num_stencil_elements; 
 	// These are pointers to gpu memory. We will need to allocate in the constructor and 
 	// copy to/from the memory in computeDeriv
 	cl::Buffer gpu_stencils;
